+++ R/showRepo.R
@@ -54,7 +54,7 @@
 #' 
 #' @note
 #' If \code{repo} and \code{user} are set to \code{NULL} (as default) in the Github mode then global parameters
-#' set in \link{setGithubRepo} function are used.
+#' set in \link{setGithubRepo} (or via \link{aoptions}) function are used.
 #' 
 #' @author 
 #' Marcin Kosinski , \email{m.p.kosinski@@gmail.com}