+++ NEWS.md
@@ -3,7 +3,8 @@ archivist 1.9
 * **New functions**:
     1. `alink` function: Returns a Link To Download an Artifact Stored on GitHub Repository. Ideal combination with `archive`
     2. Created wrappers around existing local and GitHub functions: `multiSearchInRepo`, `loadFromRepo`, `deleteRepo`
-        1. Provided new functions `deleteLocalRepo` (previous `deleteRepo`) and `deleteGithubRepo`. `deleteRepo` is now a wrapper around those both. [[#156](https://github.com/pbiecek/archivist/issues/156)] 
+        1. Provided new functions `deleteLocalRepo` (previous `deleteRepo`) and `deleteGithubRepo`. `deleteRepo` is now a wrapper around those both. [[#156](https://github.com/pbiecek/archivist/issues/156)].
+    3. `createGithubMDGallery` that give the markdown summary for each artifact in the repository. Ideal for README.md file. Example [[#144](https://github.com/pbiecek/archivist/issues/144#issuecomment-174192366)]
 * **Bugs fixed**:
     1. `asearch` function enables a user to read artifacts from default GitHub repository. In the previous version it was possible only in default local repository.
     2. It is now possible to unset global Repository with `apotions('repo/repoDir', NULL, unset = TRUE)` [[#176](https://github.com/pbiecek/archivist/issues/176)].